requirements.txt
asgiref==3.10.0
certifi==2025.11.12
click==8.3.0
coverage==7.13.1
dj-database-url==3.0.1
Django==5.2.7
django-health-check==3.20.0
django-jazzmin==3.0.1
django-ratelimit==4.1.0
django-recaptcha==4.1.0
django-redis==6.0.0
django-tenants==3.9.0
factory_boy==3.3.3
Faker==40.1.2
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
qrcode==8.2
redis==7.1.0
sentry-sdk==2.45.0
sqlparse==0.5.3
urllib3==2.5.0
uvicorn==0.38.0
whitenoise==6.11.0
